refactor: use correct path for libs
